Code Cleanup and refactor
